% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignForceZ.R
\name{AlignForceZ}
\alias{AlignForceZ}
\title{Perform a baseline correction to an AFM F-z curve}
\usage{
AlignForceZ(Z, Force, BR)
}
\arguments{
\item{BR}{minimum detach point in the retract segment of the curve
that defines the bulk region}

\item{(Z, }{Force) reversely indexed}
}
\value{
aligned force and zero point
}
\description{
This function performs the baseline correction to an AFM F-z curve
using the minimum rupture point, higher index closer to surface
}
\examples{

}
