% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootGamma.R
\name{bootGamma}
\alias{bootGamma}
\title{Estimate Free Energy Error Using Bootstrap for Gamma}
\usage{
bootGamma(data, replicates = 1000)
}
\arguments{
\item{data}{external work in kT.}
}
\description{
This function predicts mean and sd of free energy.
}
\keyword{gamma}
