% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootNorm.R
\name{bootNorm}
\alias{bootNorm}
\title{Estimate Free Energy Error Using Bootstrap for Norm}
\usage{
bootNorm(data, replicates = 1000)
}
\arguments{
\item{data}{external work in kT.}
}
\description{
This function predicts mean and sd of free energy.
}
\keyword{norm}
