% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSlope.R
\name{CalcSlope}
\alias{CalcSlope}
\title{Map F-D curves to Derivative Space}
\usage{
CalcSlope(X, width)
}
\arguments{
\item{X}{a matrix on the form [1 z Force],}

\item{width}{radius of the window for the local regression (in vector position units)}
}
\value{
A vector containing the derivatives.
}
\description{
The derivative are approximated by the slope of a linear fit to the data in a window of
a given radius width.
}
\examples{
n <- 100
x <- seq(0, 2 * pi, length.out = n)
y = sin(x) + 0.1 * rnorm(n)
X <- matrix(c(rep(1, n), x, y), nrow = n, ncol = 3)
width <- 5
b <- CalcSlope(X, width)
plot(x[(width + 1) : (n - width)], b, xlab = "x", ylab = "y", type = "l")
lines(x, y, col = "red")
legend("bottomleft", c("Slopes","Signal"), col = c(1, 2), lty = 1)
}
