% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindBPoints.R
\name{FindBPoints}
\alias{FindBPoints}
\title{Locate significant minima on an AFM F-Z curve}
\usage{
FindBPoints(Force, jDP, width, noise)
}
\arguments{
\item{Force}{higher index closer to surface}

\item{jDP}{discontinuous points from function FindDPoints}

\item{width}{radius of a window}

\item{noise}{signal threshold}
}
\value{
jBP vector of binding points
}
\description{
This function locates significant minima on an AFM F-Z curve
}
