% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDPoints.R
\name{FindDPoints}
\alias{FindDPoints}
\title{Locate discontinuous points on an AFM F-Z curve}
\usage{
FindDPoints(Z, Force, Alpha, LOD)
}
\arguments{
\item{(Z, }{Force) reversely indexed}
}
\value{
jDP vector of discontinuous points
}
\description{
This function locates discontinuous points on an AFM F-Z curve
discontinuous points correspond to positive peaks of the derivatives
}
\examples{

}
