% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindKeyPoints.R
\name{FindKeyPoints}
\alias{FindKeyPoints}
\title{Find binding points, zero force points, quasi-equilibrium points, do alignment}
\usage{
FindKeyPoints(myData, Alpha, LOD, mulBase)
}
\arguments{
\item{Alpha}{percent of points in bulk}

\item{LOD}{noise level in sigma unit}

\item{mulBase}{baseline perturbation level}
}
\value{
jDP, jZP, jBP, jQP and aligned force
}
\description{
This function locates binding points, zero force points, quasi-equilibrium points
and performs the baseline correction on an AFM F-Z curve
}
\examples{

}
