% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMinima.R
\name{FindMinima}
\alias{FindMinima}
\title{Locate all minima on an AFM F-Z curve}
\usage{
FindBPoints(Force, jDP, width)
}
\arguments{
\item{Force}{higher index closer to surface}

\item{jDP}{discontinuous points from function FindDPoints}

\item{width}{radius of a window}
}
\value{
jBP vector of minimum points
}
\description{
This function locates all minima on an AFM F-Z curve
}
