% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindQPoints.R
\name{FindQPoints}
\alias{FindQPoints}
\title{Locate quasi-equilibrium points on an AFM F-Z curve}
\usage{
FindQPoints(Z, Force, jBP, noise)
}
\arguments{
\item{jBP}{index corresponding to significant force minima}

\item{noise}{lower bound of baseline}

\item{(Z, }{Force) retraction force high index close to surface}
}
\value{
jQP vector of quasipoints
}
\description{
This function locates quasi-equilibrium points before pulling
based on zero force criterion within error bar
}
