% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitWLC.R
\name{FitWLC}
\alias{FitWLC}
\title{Fit into Worm-Like Chain Model}
\usage{
FitWLC(myData, nlsEr = 4)
}
\arguments{
\item{myData}{stored jBP, jQP, Z, Q and Force}

\item{nlsEr}{tolerence for fitting error}
}
\value{
iBP, iQP, iCP, WLC
}
\description{
This function fits force-separation curves into worm-like chain
}
\examples{

}
