% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaEstimator.R
\name{GammaEstimator}
\alias{GammaEstimator}
\title{Estimate Free Energy Assuming Underlying Gamma Distribution}
\usage{
GammaEstimator(wrkDir, data)
}
\arguments{
\item{wrkDir}{Directory for image output.}
}
\description{
This function tests underlying gamma distribution of work and estimates free energy.
}
\keyword{gamma}
