% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianEstimator.R
\name{GaussianEstimator}
\alias{GaussianEstimator}
\title{Estimate Free Energy Assuming Underlying Norm Distribution}
\usage{
GaussianEstimator(wrkDir, data)
}
\arguments{
\item{wrkDir}{Directory for image output.}
}
\description{
This function tests underlying norm distribution of work and estimates free energy.
}
\keyword{norm}
