% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistPlot.R
\name{HistPlot}
\alias{HistPlot}
\title{Plot histograms of all variables}
\usage{
HistPlot(outDir, Temp, springK)
}
\arguments{
\item{outDir}{Directory is where results.RData locates.}

\item{Temp}{temperature in Kelvin}
}
\description{
This function extracts all parameters and do histogram plot
}
