% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDirNoPeaks.R
\name{MakeDirNoPeaks}
\alias{MakeDirNoPeaks}
\title{Make subdirectories in output directory according to number of peaks}
\usage{
MakeDirNoPeaks(desDir, n = 4)
}
\arguments{
\item{desDir}{output directory}
}
\description{
Make subdirectories in output directory according to number of peaks
}
