% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcess.R
\name{PostProcess}
\alias{PostProcess}
\title{Get rid of outliers}
\usage{
PostProcess(outDir, QL = 0, QR = 200, FL = -1000, FR = 0, WL = 0,
  WR = 1000)
}
\arguments{
\item{outDir}{Directory is where work.RData locates.}
}
\description{
This function deletes all outliers
}
