% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TiltCorrection.R
\name{TiltCorrection}
\alias{TiltCorrection}
\title{Perform a baseline correction to an AFM F-Z curve}
\usage{
TiltCorrection(Z, Force, Alpha)
}
\arguments{
\item{Alpha}{percent of points in bulk}

\item{(Z, }{Force) reversely indexed}
}
\value{
aligned force and zero point
}
\description{
This function performs the baseline correction to an AFM F-Z curve
setting force at points in bulk to be zero
}
\examples{

}
