% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrapZ.R
\name{TrapZ}
\alias{TrapZ}
\title{Simple Integration}
\usage{
TrapZ(x, y)
}
\arguments{
\item{x}{Z coordinate in nm}

\item{y}{Force values in pN}
}
\value{
work in pNnm
}
\description{
This function estimates integration
}
\examples{

}
